QA Chat with Website Information-Google Gemini/requirements.txt
streamlit
google-generativeai
python-dotenv
faiss-cpu
langchain
langchain-community
langchain_google_genai